+++ R/find_map_gaps.R
@@ -11,6 +11,8 @@
 #' numeric index of marker to left, marker to right of gap, numeric
 #' index of marker to right, and the length of the gap.
 #'
+#' @seealso \code{\link{reduce_map_gaps}}
+#'
 #' @examples
 #' iron <- read_cross2(system.file("extdata", "iron.zip", package="qtl2geno"))
 #' find_map_gaps(iron$gmap, 40)